型を複数ファイルに分割
