+++ src/types/blog.ts
@@ -1,0 +1,8 @@
+import { MicroCMSListResponse } from "microcms-js-sdk";
+
+export type Blog = {
+  title: string;
+  body: string;
+};
+
+export type BlogData = MicroCMSListResponse<Blog>;
